Avoid using the glyph arrangement cache to fix font issues (#334)
(it even seems to work faster now too?)

+++ Source/UI/Sequencer/PatternRoll/ClipComponent.h
@@ -48,7 +48,7 @@ public:
     inline const Colour &getEventColour() const noexcept
     { return this->eventColour; }
 
-    Rectangle<int> getTextArea() const noexcept;
+    Rectangle<float> getTextArea() const noexcept;
 
     //===------------------------------------------------------------------===//
     // MidiEventComponent
